rds-info: Fix congestion field in rds-info -k
Orabug: 36349486

With commit 464bfeef ("rds-info: Userspace change to display congested
socket on rds-info -k") and its kernel part, we received -1 for 'no
congestion' and 1 for 'congestion'. We used to remap the 'no congestion'
to 0 before printing. This was done to work with older kernel which
doesn't send the congestion value to userspace. Now we can handle the
unsupported kernels in user space, by initializing the userspace value
to -1. If older kernel doesn't fill this value, it remains -1 in
userspace.

With this change, we would print -1 for 'no congestion' for few kernels
where commit 978f8509 ("rds: Include congested flag in rds_sock
struct.") is present. In such cases, we have to manually examine if -1
means not supported kernel or no congestion. Going forward, 0 would mean
no congestion, 1 would mean congestion and -1 would mean non-supported
kernel.

Fixes: 464bfeef ("rds-info: Userspace change to display congested socket
on rds-info -k")
Signed-off-by: Juan Garcia <juan.ma.garcia@oracle.com>
Reviewed-by: Mark Haywood <mark.haywood@oracle.com>