rds-tools: adding rds unit tests to rds tools
Orabug: 36464495

There is a need to maintain the unit tests in the rds-tools
distribution. All the unit test code goes into the /tests
subdirectory inside rds-tools. This code is delivered to the
end customer in the form of actual code along with a Makefile.

When the rds-tools package is installed by rpm(8), the unit
test code will be installed under the rds-tools documentation
directory.

Signed-off-by: Devesh Sharma <devesh.s.sharma@oracle.com>
Reviewed-by: Mark Haywood <mark.haywood@oracle.com>