rds-tools: Run autoconf as part of rpm build
Orabug: 36622987

rds-tools.spec should be running autoconf as part of the rpm build to
process the configure.in file. This will fix a bug where the usage
output from the tools display old version numbers. This fix will
remove the prebuilt configure file that was added to the project many
years ago and has not been modified since.

Reported-by: Devesh Sharma <devesh.s.sharma@oracle.com>
Signed-off-by: Mark Haywood <mark.haywood@oracle.com>
Signed-off-by: Devesh Sharma <devesh.s.sharma@oracle.com>
Reviewed-by: Devesh Sharma <devesh.s.sharma@oracle.com>