rds-tools: Conditionalize some rds.h updates
rds.h was updated by 815efae5e40a ("rds-tools: rds.h updated"). Several
of the updates were not supported upstream. This patch conditionally
adds the changes that are not supported upstream. This requires
changes to rds-stress.c and rds-info.c to conform with what upstream
supports.

Make the support conditional for oracle.

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>